Policies, Images and Validations

+++ app/Http/Controllers/PhoneController.php
@@ -4,27 +4,34 @@ namespace App\Http\Controllers;
 
 use App\Models\Phone;
 use App\Models\Color;
-use App\Models\Images;
+use App\Models\Image;
 use App\Models\Manufacturer;
 use App\Models\Operative_System;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use App\Models\ImagePhone;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Gate;
+
 
 
 class PhoneController extends Controller
 {
     private $reglasValidacion = [
-        'manufacter' => 'required|numeric|exists:manufacters', //Posiblemente se necesite el ID aquí
-        'color' => 'required|numeric|exists:colors',
-        'os' => 'required|numeric|exists:operative_systems',
+        'manufacturer' => 'required|numeric|exists:manufacturers,id', //Posiblemente se necesite el ID aquí
+        'color' => 'required|numeric|exists:colors,id',
+        'os' => 'required|numeric|exists:operative_systems,id',
         'model' => 'required|min:5|max:45',
         'name' => 'required|min:5|max:45',
         'storage' => 'required|numeric|in:64,128,256,512,1024',
         'ram' => 'required|numeric|max:12',
-        'batery' => 'required|numeric|max:7000',
-        '5g_capable' => 'required|boolean',
+        'battery' => 'required|numeric|min:2800|max:7000',
+        'fiveg_capable' => 'required|boolean',
         'release_year' => 'required|numeric|min:2015|max:2022',
         'price' => 'required|numeric|min:0',
+        'images.*' => 'mimes:jpeg,png,jpg,gif,svg|max:2048',
     ];
     
     /**
@@ -35,21 +42,24 @@ class PhoneController extends Controller
     public function index()
     {
         $phones = Phone::all(); //or get
-        return view('home', compact('phones'));
+        $images = Image::all();
+        return view('home', compact('phones', 'images'));
     }
 
     public function controlPanel()
     {
-        $phones = Phone::all(); //or get
-
-        $images = Images::all(); ///Esto se debe arreglar
+        if (!Gate::allows('hasPermission')) {
+            abort(403);
+        }
+        $phones = Phone::all();
+        $images = Image::all();
         return view('admin.tablePhones', compact('phones','images'));
     }
     
     public function myOrders()
     {
         $orders = Auth::user()->order('user')->get(); //listado de solo ese usuario 
-        return view('phones.orders', compact('orders'));
+        return view('orders.indexOrders', compact('orders'));
     }
 
     /**
@@ -59,13 +69,15 @@ class PhoneController extends Controller
      */
     public function create()
     {
+        if (!Gate::allows('hasPermission')) {
+            abort(403);
+        }
         $os = Operative_System::all();
         $manufacturers = Manufacturer::all();
         $colors = Color::all();
-
         return view('phones.formPhones',  compact('os', 'manufacturers', 'colors'));
     }
-
+    
     /**
      * Store a newly created resource in storage.
      *
@@ -74,27 +86,41 @@ class PhoneController extends Controller
      */
     public function store(Request $request)
     {
+        if (!Gate::allows('hasPermission')) {
+            abort(403);
+        }
         $request ->validate($this->reglasValidacion);
-
-        //dd('accede a metodo store');
-        //dd($request->all());
-   
+        
         $phone = new Phone();
-        $phone ->manufacter = $request ->manufacter;
-        $phone ->color = $request ->color;
-        $phone ->os = $request ->os;
+        $phone ->manufacturer_id = $request ->manufacturer;
+        $phone ->color_id = $request ->color;
+        $phone ->os_id = $request ->os;
         $phone ->model = $request ->model;
         $phone ->name = $request ->name;
         $phone ->storage = $request ->storage;
         $phone ->ram = $request ->ram;
-        $phone ->batery = $request ->batery;
-        //$phone ->5g_capable = $request ->5g_capable;
+        $phone ->battery = $request ->battery;
+        $phone ->fiveg_capable = $request ->fiveg_capable;
         $phone ->release_year = $request ->release_year;
         $phone ->price = $request ->price;
-        $phone ->photo = $request ->photo;
-        $phone ->save();
+        $phone->save();
 
-        return redirect('/phone');
+        if($request->images){
+            foreach($request->images as $key=> $image)
+            {
+                $img= new Image();
+                $imageName = $request['model'].'-img-'.time().rand(1,1000).'.'.$image->extension();
+                $img ->image = $imageName;
+                $image->move(\public_path("/phones"), $imageName);
+                $img->save();
+                $imgPhone = new ImagePhone();
+                $imgPhone->phone_id = $phone->id;
+                $imgPhone->image_id = $img->id;
+                $imgPhone->save();
+            }
+        }
+
+        return redirect('/ControlPanel')->with('success', $phone->name.' created successfully.');
         
     }
 
@@ -106,7 +132,9 @@ class PhoneController extends Controller
      */
     public function show(Phone $phone)
     {
-        return view('phones.showPhone', compact('phone'));
+        $img = ImagePhone::where('phone_id', $phone->id)->get('image_id');
+        $images= Image::whereIn('id', $img)->get();
+        return view('phones.showPhones', compact('phone', 'images'));
     }
 
     /**
@@ -117,7 +145,15 @@ class PhoneController extends Controller
      */
     public function edit(Phone $phone)
     {
-        return view('phones.formPhones', compact('phone'));
+        if (!Gate::allows('hasPermission')) {
+            abort(403);
+        }
+        $os = Operative_System::all();
+        $manufacturers = Manufacturer::all();
+        $colors = Color::all();
+        $img = ImagePhone::where('phone_id', $phone->id)->get('image_id');
+        $images= Image::whereIn('id', $img)->get();
+        return view('phones.formPhones',  compact('os', 'manufacturers', 'colors', 'phone', 'images'));
     }
 
     /**
@@ -129,23 +165,45 @@ class PhoneController extends Controller
      */
     public function update(Request $request, Phone $phone)
     {
+        if (!Gate::allows('hasPermission')) {
+            abort(403);
+        }
+
         $request ->validate($this->reglasValidacion);
 
-        $phone ->manufacter = $request ->manufacter;
-        $phone ->color = $request ->color;
-        $phone ->os = $request ->os;
+        $phone = new Phone();
+        $phone ->manufacturer_id = $request ->manufacturer;
+        $phone ->color_id = $request ->color;
+        $phone ->os_id = $request ->os;
         $phone ->model = $request ->model;
         $phone ->name = $request ->name;
         $phone ->storage = $request ->storage;
         $phone ->ram = $request ->ram;
-        $phone ->batery = $request ->batery;
-        //$phone ->5g_capable = $request ->5g_capable;
+        $phone ->battery = $request ->battery;
+        $phone ->fiveg_capable = $request ->fiveg_capable;
         $phone ->release_year = $request ->release_year;
         $phone ->price = $request ->price;
-        $phone ->photo = $request ->photo;
-        $phone ->save();
+        $phone->save();
 
-        return redirect('/phone/' . $phone->id);
+        if($request->images){
+            foreach($request->images as $key=> $image)
+            {
+                if(!property_exists($image, 'image'))
+                {
+                    $img= new Image();
+                    $imageName = $request['model'].'-img-'.time().rand(1,1000).'.'.$image->extension();
+                    $img ->image = $imageName;
+                    $image->move(\public_path("/phones"), $imageName);
+                    $img->save();
+                    $imgPhone = new ImagePhone();
+                    $imgPhone->phone_id = $phone->id;
+                    $imgPhone->image_id = $img->id;
+                    $imgPhone->save();
+                }
+            }
+        }
+
+        return redirect('/ControlPanel')->with('success', $phone->name.' updated successfully!');
     }
 
     /**
@@ -156,12 +214,33 @@ class PhoneController extends Controller
      */
     public function destroy(Phone $phone)
     {
+        if (!Gate::allows('hasPermission')) {
+            abort(403);
+        }
+        $list = ImagePhone::where('phone_id', $phone->id)->get('image_id');
+        $images= Image::whereIn('id', $img)->get();
+        foreach($images as $image)
+        {
+            if(File::exists("phones/".$image->image))
+            {
+                File::delete("images/".$image->image);
+            }
+        }
         $phone ->delete();
-        return redirect('/ControlPanel')->with('success', 'Phone Deleted successfully.');
+        return redirect('/ControlPanel')->with('success', $phone->name.' deleted successfully!');
     }
 
-    public function bindImages(Request $request)
+    public function deletePhoto($id)
     {
-
+        if (!Gate::allows('hasPermission')) {
+            abort(403);
+        }
+        $img = Image::where('id', $id)->get()->first();
+        if(File::exists("phones/".$img->image))
+        {
+            File::delete("images/".$img->image);
+        }
+        $img->delete();
+        return back()->with('success', 'Image deleted successfully!');
     }
 }
